Significant changes to spectrometer operation
Spectrometer template no longer depends on a communications protocol.
Seabreeze instead is a subclass of Instrument. A QThread is used to
acquire spectra data in the GUI. A regular Thread also works with code
commented out. Issues occur in both when the acquisition rate is faster
than the GUI can handle. Fixed refresh rates are implemented to fix
this. These can be changed after a display class instance is created.
Properties are more clearly defined to allow use of get methods. Use of
the spectrometers class with a single spectrometer no longer crashes
the GUI. Seabreeze should now raise an error if testing with no
spectrometers detected.