Fixed metadata handling in Camera
I’d updated the way cameras handle parameters, but not the code that
retrieves them for metadata.  This is now done - I think it’s much
simpler.  Essentially, at initialisation time, we populate
self.metadata_property_names with the names of the camera’s parameters
(in addition to anything inherited from Instrument).  That removes the
need for anything camera-specific in get_metadata.