Various bug-fixes
Fixed an issue where the info renderer could crash if run on a
DummyHDF5Group.

Fixed an issue where 1D PyQtGraphs failed on groups if the first item
was a 1D dataset but subsequent ones were not.

Removed the minimum size for 2D PyQtGraphs.  If people need a minimum
(or default) size for the data browser, this should be set in the data
browser…