progress bar - fix/blue tack solution
The progress bar has been forcing other apps to freeze upon completion, this is because it exec's its only gui eventloop. Therefore I have removed the command, this is currently not really a fix and it prevents the progress bar from blocking access to the gui's that are currently running. This is hoever good enough for our uses, where we trust are users not to do anything stupid - but a better fix based around the progress bar modality would be a cleaner solution