Update to experimental triax code
In current triax code, the relationship between the pixel any wavelength lands on and stepper motor postions is approximated as quadratic. The realtionship between these quadratric coefficents and wavelengths is also taken as quadratic so that the system can be inverted. This is not ideal. The new system uses this as an intial guess, then iteratativly updates the wavelengths so that the realtionship is represented by a quadratic spline instead. Also now does not re-calculate wavelength array every time is it requested.