Moved SERS fitting modules and added new one.
Move modules to sub-folder to reduce clutter. Wrote Iterative_Raman_Fitting.py, which auto-fits raman on a zero background. This runs more slowly than Auto-Fit_Raman, but does not rely on finding local maxima. It instead starts with no peaks and iterativly attempts to add a new peak in a locally optimal position.