Added Stokes_AntiStokes_Fitting.py - auto-fits a SERS signal taken over the notch filter
This code fits polynomial backgrounds either side of the laser, fits peaks as Lorentzian peaks and models the transmission of the notch filter as two sigmoid functions. This does not require the user to submit any more information then the signal and the Raman Shift axis. Useful mainly for removing background close to the laser line while being mindful not to move the peaks around.