Slight update of old code, and a bunch of new modules
Updated Condense_DF_Spectra to increase robustness, such as it is.
Updated spc_to_h5 to give more sensible h5 dataset names, and also to return the h5 filename when running the main function.

Added a bunch of new modules containing classes for the various types of spectra we collect with simple (and not-so-simple) plotting functions for each.

Most of the standalone functions in the spectrum_tools module are commented and ready-to-use. Please try them out and let me know any bugs!

The dft_raman_tools module is based on an old version of David's code; it runs but is a bit slow and the inner workings are a bit convoluted. However, the master control and DFT collection objects should hopefully work as easy wrappers for it.

npom_df_pl_tools will eventually replace DF_Multipeakfit and Condense_Fit_DF. It's a (big) work in progress but should be available soon and much more user-friendly than the old code.

All these new modules are still beta versions. I wrote them mostly for myself but sharing them here in case others find them useful.
Please feel free to use and copy any of the modules and classes, but don't modify yet as I'm still actively working on them.